add Match parser, fix some code style

+++ Dota2Analytics.Data/Entities/Player.cs
@@ -7,7 +7,7 @@ namespace Dota2Analytics.Data.Entities
         public string? Name { get; set; }
         public bool IsPrivateHistory { get; set; }
         public string NickName { get; set; }
-        public int? SteamAccountId { get; set; }
+        public long? SteamAccountId { get; set; }
         public int WinRate { get; set; }
         public int Rank { get; set; }
         public int? MatchEventId { get; set; }
